Create a website to show movie review for favorite films
This website has a main page and links to individual movie reviews.
At any time you can click on "Julio's Movie Reviews" to go back to
the home page. Clicking on a movie image or text will take you to a
review where you can see a rating, additional images, and some
analysis of the film.